Features changed (Tests): Changed testing features and docs
* TEST_SET: Changed the "nanp_random_tests" to "nanp_selective_tests",
soon realized that the testing is rather poor as is the design. This
thinking traces back to Michael DeVane's note on last Saturday morning
that what we really want are three new classes for area-code,
central-office, and subscriber number, which opens the way to rather
strict testing of rules in those limited spaces, which is preferable to
how this test is presently constructed.